# server/models/schema_Fastcontent_db.js
/* 
* Generated by
* 
*      _____ _          __  __      _     _
*     / ____| |        / _|/ _|    | |   | |
*    | (___ | | ____ _| |_| |_ ___ | | __| | ___ _ __
*     \___ \| |/ / _` |  _|  _/ _ \| |/ _` |/ _ \ '__|
*     ____) |   < (_| | | | || (_) | | (_| |  __/ |
*    |_____/|_|\_\__,_|_| |_| \___/|_|\__,_|\___|_|
*
* The code generator that works in many programming languages
*
*			https://www.skaffolder.com
*
*
* You can generate the code from the command-line
*       https://npmjs.com/package/skaffolder-cli
*
*       npm install -g skaffodler-cli
*
*   *   *   *   *   *   *   *   *   *   *   *   *   *   *   *
*
* To remove this comment please upgrade your plan here: 
*      https://app.skaffolder.com/#!/upgrade
*
* Or get up to 70% discount sharing your unique link:
*       https://app.skaffolder.com/#!/register?friend=5e6a85b49f48c9443d27400c
*
* You will get 10% discount for each one of your friends
* 
*/
// Import Sequelize
import Sequelize from "sequelize";
import Database from "../classes/Database_Fastcontent_db";

export default init => {
  let sequelize = Database.getConnection();


    /**
      * ------------------------------------
      * Start define generated schema
      *
      * The content of this section will be overwritten on each documentation, 
      * please insert your customization at the top or at the end of the file.
      * ------------------------------------
      */



    /**
      * ------------------------------------
      * Project
      * ------------------------------------
      */
    class Project extends Sequelize.Model{}
    Project.init({
      _id: { 
        type: Sequelize.INTEGER,
        primaryKey: true,
        autoIncrement: true
      },
      
      category: {
        type: Sequelize.STRING, 
        allowNull: false
      },
      
      creation_date: {
        type: Sequelize.DATE, 
        allowNull: false
      },
      
      deadline: {
        type: Sequelize.DATE, 
        allowNull: false
      },
      
      description: {
        type: Sequelize.STRING
      },
      
      owner: {
        type: Sequelize.STRING, 
        allowNull: false
      },
      
      status: {
        type: Sequelize.STRING, 
        allowNull: false
      },
      
      timezone: {
        type: Sequelize.STRING
      },
      
      title: {
        type: Sequelize.STRING, 
        allowNull: false
      },
      
      type: {
        type: Sequelize.STRING, 
        allowNull: false
      },
      
      //RELATIONS
        
      owner:  {
        type: Sequelize.INTEGER,
        references: {
          model: "User",
          key: '_id',
        },
      },
        
      
      
      //EXTERNAL RELATIONS
      /*
      project_owner: {
        type: Sequelize.INTEGER,
        references: {
          model: User,
          key: '_id',
        }
      },
      */
    },
      { sequelize, tableName: "project", timestamps: false }
    );



    /**
      * ------------------------------------
      * User
      * ------------------------------------
      */
    class User extends Sequelize.Model{}
    User.init({
      _id: { 
        type: Sequelize.INTEGER,
        primaryKey: true,
        autoIncrement: true
      },
      
      company: {
        type: Sequelize.STRING
      },
      
      country: {
        type: Sequelize.STRING
      },
      
      industry: {
        type: Sequelize.STRING
      },
      
      mail: {
        type: Sequelize.STRING, 
        allowNull: false
      },
      
      mobile: {
        type: Sequelize.FLOAT
      },
      
      name: {
        type: Sequelize.STRING
      },
      
      password: {
        type: Sequelize.STRING, 
        allowNull: false
      },
      
      role: {
        type: Sequelize.STRING
      },
      
      surname: {
        type: Sequelize.STRING
      },
      
      user_roles: {
        type: Sequelize.STRING
      },
      
      username: {
        type: Sequelize.STRING, 
        allowNull: false
      },
      
      website: {
        type: Sequelize.STRING
      },
      
      //RELATIONS
        
        
      project_owner:  {
        type: Sequelize.INTEGER,
        references: {
          model: "Project",
          key: '_id',
        },
      },
      
      
      //EXTERNAL RELATIONS
      /*
      owner: {
        type: Sequelize.INTEGER,
        references: {
          model: Project,
          key: '_id',
        }
      },
      */
    },
      { sequelize, tableName: "user", timestamps: false }
    );


    /**
      * ------------------------------------
      * Relations many to many
      * ------------------------------------
      */

    
    
  /**
   * ------------------------------------
   * End define generated schema
      * ------------------------------------
      */

    /**
      * ------------------------------------
      * Roles
      * ------------------------------------
      */
    class Roles extends Sequelize.Model{}
    Roles.init({
      _id: { 
        type: Sequelize.INTEGER,
        primaryKey: true,
        autoIncrement: true
      },
      
      role: {
        type: Sequelize.STRING
      },
      
      //RELATIONS
        
      _user:  {
        type: Sequelize.INTEGER,
        references: {
          model: "User",
          key: '_id',
        },
      }
      
      
      //EXTERNAL RELATIONS
      /*
      */
    },
      { sequelize, tableName: "roles", timestamps: false }
    );

    User.hasMany(Roles, {
      foreignKey: "_user"
    });

    /**
      * ------------------------------------
      * Insert here your custom models
      * ------------------------------------
      */

};